requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
edge-tts>=7.2.8,<8.0.0
httpx>=0.27.0,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.0.1,<6.0.0
psycopg[binary]>=3.2.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
email-validator>=2.2.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
